+++ src/Model/Parameter/AlertDestination/HttpParameters.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Model\Parameters\AlertDestination;
+namespace App\Model\Parameter\AlertDestination;
 
-use App\Model\Parameters\DynamicParameters;
-use App\Model\Parameters\DynamicParametersInterface;
+use App\Model\Parameter\DynamicParameters;
+use App\Model\Parameter\DynamicParametersInterface;
 
 class HttpParameters extends DynamicParameters
 {
