+++ src/Model/Parameter/DynamicParameters.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Model\Parameters;
+namespace App\Model\Parameter;
 
 abstract class DynamicParameters implements DynamicParametersInterface, \ArrayAccess
 {
