+++ src/Model/Parameter/NullParameters.php
@@ -1,13 +1,15 @@
 <?php
+
 declare(strict_types=1);
 
-namespace App\Model\Parameters\AlertDestination;
+namespace App\Model\Parameter;
 
-use App\Model\Parameters\DynamicParameters;
-use App\Model\Parameters\DynamicParametersInterface;
-
-class MonologParameters extends DynamicParameters
+class NullParameters extends DynamicParameters
 {
+    private function __construct()
+    {
+    }
+
     public static function fromJsonString(string $json): DynamicParametersInterface
     {
         return new self();
